Prompt for exact version with autocomplete.
Prompting for the exact version makes it easier to specify custom
version numbers to tag, including versions with build/prerelease data.

The simple behavior behind the common cases of patch/minor/major
increments are maintained [for the most part] by utilizing autocomplete
where the versions are automatically entered for the user as they type.