Don't die on the first commit to a repository.
This tool should be able to build release notes even for a repository
that has no prior releases.  One problem with such a repository is that
we go back to the very first commit in the repository when viewing the
history (by design).  This commit has no parents (it should be fairly
unique in this regard).

This code was expecting there to always be a parent commit and was
causing an underflow exception when trying to access the first vertex in
the set of parent vertices.  A simple check on `isEmpty` lets us handle
this case.  It's not an error case, but we do need to stop iterating
(there is no new `current` to walk to).  The last commit (or first
commit depending on your POV) will still get included it the list of
"first parents" since it may include important changes and should be
represented by the release notes.