Upgrade from the DialogHelper to the QuestionHelper.
Symphony has deprecated the DialogHelper and replaced it with a new
QuestionHelper.  This updates the code to use the newer helper without
otherwise changing behavior.

Because of how the new QuestionHelper is created, we now need to pass
the InputInterface through to the prompt.  This is the cause of most of
the changes in the code.

Because the QuestionHelper is newish, I also bumped the required
symphony/console version up to 2.7.