Change primary UI to a menu loop.
This is a major change of the UI to a loop that asks you to choose what
step to take.  You can revisit previous steps and tweak things until the
release notes are finalized.

The proposed reasons for this change:
* It was super annoying when you answered a prompt incorrectly before as
  there was no way to go back and fix it without restarting from the
  beginning.
* Adding new features with new prompts was becoming very difficult and
  made the tool even harder to use with each new prompt.  By pulling
  things down into a simple loop, adding new choices or beefing out the
  existing choices is made easier.
* There were a lot of prompts in the old UI that you generally went with
  the default answer on.  These prompts can be skipped entirely now.

The diff for this is nasty, as is some of the code.  It was difficult to
find a way of introducing this change without making a massive amount of
changes, but I tried to pull out as much as I could.